refactor(kube): system resources settings
