+++ app/react/kubernetes/datatables/SystemResourceDescription.tsx
@@ -1,3 +1,5 @@
+import { Authorized } from '@/react/hooks/useUser';
+
 import { TextTip } from '@@/Tip/TextTip';
 
 interface Props {
@@ -6,12 +8,15 @@ interface Props {
 
 export function SystemResourceDescription({ showSystemResources }: Props) {
   return (
-    <div className="w-full">
-      {!showSystemResources && (
-        <TextTip color="blue" className="!mb-0">
-          System resources are hidden, this can be changed in the table settings
-        </TextTip>
-      )}
-    </div>
+    <Authorized authorizations="K8sAccessSystemNamespaces" adminOnlyCE>
+      <div className="w-full">
+        {!showSystemResources && (
+          <TextTip color="blue" className="!mb-0">
+            System resources are hidden, this can be changed in the table
+            settings
+          </TextTip>
+        )}
+      </div>
+    </Authorized>
   );
 }
