+++ app/react/kubernetes/ingresses/IngressDatatable/IngressDatatable.tsx
@@ -75,17 +75,12 @@ export function IngressDatatable() {
       renderTableActions={tableActions}
       renderTableSettings={() => (
         <TableSettingsMenu>
-          <DefaultDatatableSettings
-            settings={tableState}
-            hideShowSystemResources={!canAccessSystemResources}
-          />
+          <DefaultDatatableSettings settings={tableState} />
         </TableSettingsMenu>
       )}
       description={
         <SystemResourceDescription
-          showSystemResources={
-            tableState.showSystemResources || !canAccessSystemResources
-          }
+          showSystemResources={tableState.showSystemResources}
         />
       }
       disableSelect={useCheckboxes()}
