+++ app/react/kubernetes/services/ServicesView/ServicesDatatable/datatable-store.ts
@@ -1,9 +1,7 @@
-import { refreshableSettings, createPersistedStore } from '@@/datatables/types';
+import { TableSettings } from '@/react/kubernetes/datatables/DefaultDatatableSettings';
+import { systemResourcesSettings } from '@/react/kubernetes/datatables/SystemResourcesSettings';
 
-import {
-  systemResourcesSettings,
-  TableSettings,
-} from '../../../datatables/DefaultDatatableSettings';
+import { refreshableSettings, createPersistedStore } from '@@/datatables/types';
 
 export function createStore(storageKey: string) {
   return createPersistedStore<TableSettings>(storageKey, 'name', (set) => ({
